website and webmail in different servers
I understand that using example.com/mail for webmail implies that example.com must be in same server. This is not the case if website is in S3/CloudFront/LoadBalancing or another EC2 instance.
So I suggest using webmail.example.com for HTTP/HTTPS webmail access.
You will need to update Route 53 settings in tutorial.